@ refactor(866): move HumanLifecycleService into Humans.Users (G5 lane 4b-2d, part 1)
Peter, 2026-08-14: membership machinery is Users, NOT Governance —
Governance is for governance only (votes, membership decisions).

IHumanLifecycleService -> Humans.Users.Contracts (public: four consumers
outside the section — Consent, Development, Onboarding, and Users own
admin controller). HumanLifecycleService -> Humans.Users/Services,
internal sealed. Registration moves from Shell to the section Section.cs.

Zero new project references in either direction. The leaf already
referenced Humans.Onboarding.Contracts (for OnboardingResult) and
Humans.Interfaces (for IOrchestrator, which is kept per finding 53 —
the marker already ships in the right assembly). All four consumers
already imported Humans.Users.Contracts, so the move is a deleted using
line each.

Its two test files move to tests/Humans.Users.Tests; the reflection
anchor in HumanLifecycleArchitectureTests is repointed rather than left
to silently cover nothing.

Refs nobodies-collective/Humans#866

@
