+++ src/Sections/Humans.Onboarding/Docs/Onboarding.md
@@ -1,7 +1,7 @@
 <!-- freshness:triggers
   src/Sections/Humans.Onboarding/**
   src/Sections/Humans.Onboarding.Contracts/**
-  src/Humans.Application/Services/HumanLifecycle/**
+  src/Sections/Humans.Users/Services/HumanLifecycleService.cs
   src/Humans.Application/Services/Profiles/ProfileService.cs
   src/Humans.Application/Services/Users/UserService.cs
   src/Humans.Application/Services/Users/AccountProvisioningService.cs
@@ -125,7 +125,7 @@ After the nobodies-collective#584 narrowing, `OnboardingService` injects only wh
 - **67 resx keys came home.** Two callers outside the section inject `IStringLocalizer<OnboardingResource>` rather than the prefixes being split: Shell's `ShiftsController` (`Onboarding_NameRequiredBeforeShifts`) and Governance's board-voting detail page (four `OnboardingReview_*` applicant-summary labels). Three keys deliberately stayed in `SharedResource` — `Onboarding_BurnerNameLabel`, `Onboarding_LegalFirstNameLabel`, `Onboarding_LegalLastNameLabel` are `[Display]` names resolved by MVC's global `DataAnnotationLocalizerProvider`, which is bound to `SharedResource` and cannot see a section's set.
 
 - `OnboardingService` lives in `src/Sections/Humans.Onboarding/Services/OnboardingService.cs` and depends only on interfaces (plus `IClock`).
-- `HumanLifecycleService` lives in `src/Humans.Application/Services/HumanLifecycle/HumanLifecycleService.cs` and exposes `IHumanLifecycleService` — the **single entry point** for admin-initiated `SuspendAsync` / `UnsuspendAsync`, plus `RestoreConsentSuspensionAsync` (clears a consent suspension once the user is back in compliance; no-op for `AdminSuspended` users). Same orchestrator shape as `OnboardingService`: owns no tables, depends only on cross-section service interfaces (`IUserService`, `INotificationEmitter`, `INotificationAutoResolve`, `IAuditLogService`, `IHumansMetrics` — the notification pair now from `Humans.Notifications.Contracts`). Suspend / unsuspend writes flow through `IUserService.ApplyProfileOnboardingMutationAsync` (a `SetSuspension` command; admin suspension sets `AdminSuspension: true`, which classifies the user as `UserState.AdminSuspended`); unsuspend resolves the user's `AccessSuspended` notifications via `INotificationAutoResolve.ResolveBySourceAsync`. The bulk grace-period suspension path used by `SuspendNonCompliantMembersJob` continues to use `IUserService.SuspendProfilesForMissingConsentAsync` directly — that path has job-specific notification, audit, and per-team Google-removal side effects that don't fit the per-user lifecycle entry point.
+- `HumanLifecycleService` lives in `src/Sections/Humans.Users/Services/HumanLifecycleService.cs` and exposes `IHumanLifecycleService` (on `Humans.Users.Contracts`) — the **single entry point** for admin-initiated `SuspendAsync` / `UnsuspendAsync`, plus `RestoreConsentSuspensionAsync` (clears a consent suspension once the user is back in compliance; no-op for `AdminSuspended` users). Same orchestrator shape as `OnboardingService`: owns no tables, depends only on cross-section service interfaces (`IUserService`, `INotificationEmitter`, `INotificationAutoResolve`, `IAuditLogService`, `IHumansMetrics` — the notification pair now from `Humans.Notifications.Contracts`). Suspend / unsuspend writes flow through `IUserService.ApplyProfileOnboardingMutationAsync` (a `SetSuspension` command; admin suspension sets `AdminSuspension: true`, which classifies the user as `UserState.AdminSuspended`); unsuspend resolves the user's `AccessSuspended` notifications via `INotificationAutoResolve.ResolveBySourceAsync`. The bulk grace-period suspension path used by `SuspendNonCompliantMembersJob` continues to use `IUserService.SuspendProfilesForMissingConsentAsync` directly — that path has job-specific notification, audit, and per-team Google-removal side effects that don't fit the per-user lifecycle entry point.
 - **Decorator decision — no caching decorator.** Onboarding owns no cached data. Every cache invalidation for an Onboarding-driven write happens inside the owning section's service or decorator (Profile decorator refreshes `_byUserId` after clear/flag/reject/suspend/unsuspend; `INavBadgeCacheInvalidator` and `INotificationMeterCacheInvalidator` fire from the same write path; `IVotingBadgeCacheInvalidator` fires from `IApplicationDecisionService.CastBoardVoteAsync`).
 - **Cross-domain navs stripped:** N/A — Onboarding owns no entities.
 - **DI direction is one-way.** `OnboardingService → IUserService` (and other leaves) only. No leaf depends on `IOnboardingService`. The historical `IOnboardingEligibilityQuery` narrow-interface band-aid is removed; the threshold check stayed on `IOnboardingService.SetConsentCheckPendingIfEligibleAsync` (it's director-level work — predicates over Profile + Consent + Legal) and is now invoked by the controllers as a peer call to the leaf-service write. Reviewers should reject any new ctor dependency from `ProfileService` / `ConsentService` onto `IOnboardingService` (or any other director) — that's the inversion this PR removed. The cycle guard `tests/Humans.Application.Tests/Services/DependencyCycleResolutionTests.NoCircularConstructorDependencies_AcrossApplicationServices` enforces this.
