+++ src/Sections/Humans.Onboarding/Services/OnboardingService.cs
@@ -3,7 +3,6 @@ using Humans.Application.DTOs;
 using Humans.AuditLog.Contracts;
 using Humans.Consent.Contracts;
 using Humans.Governance.Contracts;
-using Humans.Application.Interfaces.HumanLifecycle;
 using Humans.Domain.Constants;
 using Humans.Domain.Enums;
 using Humans.Email.Contracts;
