+++ tests/Humans.Consent.Tests/Services/ConsentServiceTests.cs
@@ -14,7 +14,6 @@ using ConsentService = Humans.Consent.Services.ConsentService;
 using Humans.Consent.Contracts;
 using Humans.Notifications.Contracts;
 using Humans.Governance.Contracts;
-using Humans.Application.Interfaces.HumanLifecycle;
 using Humans.Domain.Enums;
 using Humans.Application.Interfaces.Users;
 using Humans.Consent.Data;
