+++ tests/Humans.Development.Tests/DevPersonaSeederTests.cs
@@ -10,7 +10,6 @@ using Humans.Camps.Contracts;
 using Humans.Consent.Contracts;
 using Humans.Governance.Contracts;
 using Humans.Application.Interfaces.GoogleIntegration;
-using Humans.Application.Interfaces.HumanLifecycle;
 using Humans.Onboarding.Contracts;
 using Humans.Users.Contracts;
 using Humans.Teams.Contracts;
